back/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra dev -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
debugpy==1.8.14
    # via fastapi-base (pyproject.toml)
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.116.0
    # via fastapi-base (pyproject.toml)
fastapi-cli==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.2
    # via fastapi-cli
filelock==3.18.0
    # via virtualenv
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.176.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via fastapi-base (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mypy==1.15.0
    # via fastapi-base (pyproject.toml)
mypy-extensions==1.1.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via pytest
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via fastapi-base (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-generativeai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyparsing==3.2.3
    # via httplib2
pytest==8.4.1
    # via fastapi-base (pyproject.toml)
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
requests==2.32.4
    # via google-api-core
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.8
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.5.1
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
ruff==0.11.7
    # via fastapi-base (pyproject.toml)
sentry-sdk==2.32.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
tqdm==4.67.1
    # via google-generativeai
typer==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   fastapi
    #   google-generativeai
    #   mypy
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
virtualenv==20.31.2
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
